심화미션: 공통로직 모듈로 분리
