--- js/account.js
@@ -1,6 +1,0 @@
-const EMAIL_REGEX = new RegExp(
-  "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$"
-);
-
-const TEST_EMAIL = "test@codeit.com";
-const TEST_PW = "codeit101";
