+++ js/account/constant.js
@@ -1,0 +1,6 @@
+export const EMAIL_REGEX = new RegExp(
+  "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$"
+);
+
+export const TEST_EMAIL = "test@codeit.com";
+export const TEST_PW = "codeit101";
