+++ signin/index.html
@@ -63,6 +63,6 @@
       </div>
     </main>
 
-    <script src="./signin.js"></script>
+    <script type="module" src="./signin.js"></script>
   </body>
 </html>
