+++ signin/signin.js
@@ -1,19 +1,9 @@
-const email = document.querySelector("#email");
-const password = document.querySelector("#password");
-const emailError = document.querySelector(".emailError");
-const passwordError = document.querySelector(".passwordError");
-
-const emailRegex = new RegExp(
-  "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$"
-);
+import { EMAIL_REGEX, TEST_EMAIL, TEST_PW } from "/js/account/constant.js";
+import { eyeOpen, eyeClose } from "/js/account/passwordIconToggle.js";
+import { email, emailError } from "/js/account/email.js";
+import { password, passwordError } from "/js/account/password.js";
 
 const form = document.querySelector("#form");
-
-const eyeOpen = document.querySelector(".eyeOpen");
-const eyeClose = document.querySelector(".eyeClose");
-
-const TEST_EMAIL = "test@codeit.com";
-const TEST_PW = "codeit101";
 
 function onSubmit(event) {
   event.preventDefault();
@@ -32,7 +22,7 @@ function emailOnFocusOut(e) {
     e.target.classList.add("inputError");
     return;
   }
-  if (!emailRegex.test(e.target.value)) {
+  if (!EMAIL_REGEX.test(e.target.value)) {
     emailError.textContent = "올바른 이메일 주소가 아닙니다.";
     e.target.classList.add("inputError");
     return;
@@ -52,32 +42,6 @@ function passwordOnFocusOut(e) {
   e.target.classList.remove("inputError");
 }
 
-function emailOnInput(e) {
-  emailError.textContent = "";
-  e.target.classList.remove("inputError");
-}
-
-function passwordOnInput(e) {
-  passwordError.textContent = "";
-  e.target.classList.remove("inputError");
-}
-
-function eyeOpenOnClick(e) {
-  e.target.classList.add("display_none");
-  eyeClose.classList.remove("display_none");
-  password.type = "password";
-}
-
-function eyeCloseOnClick(e) {
-  e.target.classList.add("display_none");
-  eyeOpen.classList.remove("display_none");
-  password.type = "text";
-}
-
 email.addEventListener("focusout", emailOnFocusOut);
 password.addEventListener("focusout", passwordOnFocusOut);
-email.addEventListener("input", emailOnInput);
-password.addEventListener("input", passwordOnInput);
-eyeOpen.addEventListener("click", eyeOpenOnClick);
-eyeClose.addEventListener("click", eyeCloseOnClick);
 form.addEventListener("submit", onSubmit);
