+++ signup/index.html
@@ -84,7 +84,6 @@
       </div>
     </main>
 
-    <script src="/js/account.js"></script>
-    <script src="./signup.js"></script>
+    <script type="module" src="./signup.js"></script>
   </body>
 </html>
