+++ signup/signup.js
@@ -1,12 +1,10 @@
-const email = document.querySelector("#email");
-const emailError = document.querySelector(".emailError");
-const password = document.querySelector("#password");
-const passwordError = document.querySelector(".passwordError");
+import { EMAIL_REGEX, TEST_EMAIL } from "/js/account/constant.js";
+import { eyeOpen, eyeClose } from "/js/account/passwordIconToggle.js";
+import { email, emailError } from "/js/account/email.js";
+import { password, passwordError } from "/js/account/password.js";
+
 const passwordConfirm = document.querySelector("#passwordConfirm");
 const passwordConfirmError = document.querySelector(".passwordConfirmError");
-
-const eyeOpen = document.querySelector(".eyeOpen");
-const eyeClose = document.querySelector(".eyeClose");
 
 const form = document.querySelector("#form");
 
@@ -31,11 +29,6 @@ function emailOnFocusOut(e) {
   e.target.classList.remove("inputError");
 }
 
-function emailOnInput(e) {
-  emailError.textContent = "";
-  e.target.classList.remove("inputError");
-}
-
 function passwordOnFocusOut(e) {
   const password = e.target.value;
   if (
@@ -48,11 +41,6 @@ function passwordOnFocusOut(e) {
     e.target.classList.add("inputError");
     return;
   }
-  passwordError.textContent = "";
-  e.target.classList.remove("inputError");
-}
-
-function passwordOnInput(e) {
   passwordError.textContent = "";
   e.target.classList.remove("inputError");
 }
@@ -80,23 +68,7 @@ function onSubmit(e) {
   location.href = "/folder";
 }
 
-function eyeOpenOnClick(e) {
-  e.target.classList.add("display_none");
-  eyeClose.classList.remove("display_none");
-  password.type = "password";
-}
-
-function eyeCloseOnClick(e) {
-  e.target.classList.add("display_none");
-  eyeOpen.classList.remove("display_none");
-  password.type = "text";
-}
-
 email.addEventListener("focusout", emailOnFocusOut);
-email.addEventListener("input", emailOnInput);
 password.addEventListener("focusout", passwordOnFocusOut);
-password.addEventListener("input", passwordOnInput);
 passwordConfirm.addEventListener("input", passwordConfirmOnInput);
-eyeOpen.addEventListener("click", eyeOpenOnClick);
-eyeClose.addEventListener("click", eyeCloseOnClick);
 form.addEventListener("submit", onSubmit);
